# app/backend/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Ticket Management System API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TMS API
  version: "1.0"
paths:
...
  /v1/enterprise/tenants:
    get:
      consumes:
      - application/json
      description: Get a list of all tenants (requires super admin permissions)
      produces:
      - application/json
      responses:
        "200":
          description: List of tenants
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or missing authentication
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List tenants
      tags:
      - Tenants
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"